Rename 404 view for consistency
